Dependency Injection Implemented with Hilt
