+++ app/src/main/java/com/josejordan/newsapp/MyApplication.kt
@@ -1,0 +1,9 @@
+package com.josejordan.newsapp
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication : Application() {
+
+}
