--- app/src/main/java/com/josejordan/newsapp/NewsAPI.kt
@@ -1,25 +1,0 @@
-package com.josejordan.newsapp
-
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface NewsAPI {
-    @GET("v2/top-headlines")
-    suspend fun getTopHeadlines(
-        @Query("country") country: String,
-        @Query("apiKey") apiKey: String
-    ): NewsResponse
-
-    companion object {
-        fun create(): NewsAPI {
-            val retrofit = Retrofit.Builder()
-                .addConverterFactory(GsonConverterFactory.create())
-                .baseUrl(Constants.BASE_URL)
-                .build()
-
-            return retrofit.create(NewsAPI::class.java)
-        }
-    }
-}
