+++ app/src/main/java/com/josejordan/newsapp/api/NewsAPI.kt
@@ -1,0 +1,15 @@
+package com.josejordan.newsapp.api
+
+import com.josejordan.newsapp.data.NewsResponse
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface NewsAPI {
+    @GET("v2/top-headlines")
+    suspend fun getTopHeadlines(
+        @Query("country") country: String,
+        @Query("apiKey") apiKey: String
+    ): NewsResponse
+
+}
+
