+++ app/src/main/java/com/josejordan/newsapp/data/Source.kt
@@ -1,4 +1,4 @@
-package com.josejordan.newsapp
+package com.josejordan.newsapp.data
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
