+++ app/src/main/java/com/josejordan/newsapp/repository/NewsRepository.kt
@@ -1,5 +1,9 @@
-package com.josejordan.newsapp
+package com.josejordan.newsapp.repository
+
+import com.josejordan.newsapp.data.NewsResponse
+
 
 interface NewsRepository {
     suspend fun getTopHeadlines(country: String, apiKey: String): NewsResponse
 }
+
