+++ build.gradle
@@ -1,7 +1,18 @@
-// Top-level build file where you can add configuration options common to all sub-projects/modules.
+buildscript {
+    dependencies {
+        classpath "com.google.dagger:hilt-android-gradle-plugin:2.44"
+    }
+}
+
+
+
 plugins {
     id 'com.android.application' version '8.0.0' apply false
     id 'com.android.library' version '8.0.0' apply false
     id 'org.jetbrains.kotlin.android' version '1.8.0' apply false
     id 'org.jetbrains.kotlin.plugin.parcelize' version '1.5.0'  apply false
+}
+
+task clean(type: Delete) {
+    delete rootProject.buildDir
 }
